requirements.txt
cloudpickle==0.2.2
cymem==1.31.2
cytoolz==0.8.2
flexmock==0.10.2
murmurhash==0.26.4
nltk==3.2.2
numpy==1.12.0
pathlib==1.0.1
plac==0.9.6
preshed==0.46.4
pyenchant==1.6.8
semver==2.7.4
six==1.10.0
spacy==1.6.0
sputnik==0.9.3
thinc==6.2.0
toolz==0.8.2
tqdm==4.11.2
ujson==1.35
